ci(docker): modify to working Dockerfile
Since everything that is not needed is already blocked off by the dockerignore file, we can safely
extend from version 16 of NodeJS and copy all the contents exposed in the context over to the Docker
container. From there, we will simply follow the standard procedure of installing all the needed
packages, build the project, and start the project.